Added GCS backend to Terraform
