+++ install/terraform/backend.tf
@@ -1,0 +1,6 @@
+terraform {
+  backend "gcs" {
+    bucket = "jkterraform"
+    prefix = "state"
+  }
+}
